+++ dashboard/src/recoil/persons.js
@@ -1,10 +1,19 @@
+import localforage from 'localforage';
 import { atom, selector } from 'recoil';
 import { capture } from '../services/sentry';
 import { organisationState } from './auth';
 
+const collectionName = 'person';
 export const personsState = atom({
-  key: 'personsState',
+  key: collectionName,
   default: [],
+  effects: [
+    ({ onSet }) => {
+      onSet(async (newValue) => {
+        await localforage.setItem(collectionName, newValue);
+      });
+    },
+  ],
 });
 
 export const customFieldsPersonsMedicalSelector = selector({
@@ -332,8 +341,6 @@ export const preparePersonForEncryption = (customFieldsMedical, customFieldsSoci
 export const commentForUpdatePerson = ({ newPerson, oldPerson }) => {
   try {
     const commentbody = {
-      type: 'person',
-      item: newPerson._id,
       person: newPerson._id,
     };
     const notifyChange = (field, before, now) => `Changement ${field}: